Add testsuite infrastructure and initial test
This adapts the test infrastructure from the "language-haskell" repo.
The use of "local -n" in the bash script has been removed, so that it
runs on older versions of bash (as found on older macOS).  It also
avoids an interactive prompt to install the "vscode-tmgrammar-test"
package in npx, instead checking for the package and installing it
if necessary and printing a note for the user.

The script supports multiple syntaxes depending on the file extension,
which was necessary for "language-haskell".  The BH repo has only
one syntax, but this feature was left in, in case it is needed in the
future.

The suite has only one test, adapted from "language-haskell".  There
is also only one test directory, but more can be added, as indicated
by commented-out lines in the shell script.