Install packages properly
Users of the repo should run "npm ci" to install the exact packages
as specified in the "package-lock.json" file.  This is particularly
true for the GitHub CI job, which should be testing that file and
be reproducable.  Thus, we change the CI job to run "npm ci".
That requires that "vscode-tmgrammar-test" be added to the list of
dependencies in "package.json" (and the lock file updated).  The
"js-yaml" package (used to generate the grammar files) was not in the
list of explicit dependencies, but it did appear in the lock file,
because it was pulled in as a dependency of a dependency; instead,
we now list it explicitly.

If the user did not install the required packages, but attempts to
run a package, the call can hang because it is interactively waiting
for the user to respond y/n to installing the missing package.  The
testsuite script attempted to handle this by responding "yes" to the
prompt, but that would not install via the lock file.  Instead, the
script now reports an error, with a message suggesting that the user
run "npm ci" or "npm install".

Similarly, the Makefile can hang when when building the grammars,
if the "js-yaml" package is not installed.  This is improved by
having the Makefile check that the required packages are installed
and, if not, report an error suggesting how to install.

There are other cleanups to the Makefile and a "clean" target added.