Fix the test script and add library snapshot tests
The attempt to support older bash was broken, as the arguments to
runTests were not being passed as array.  Newer bash can be installed
on macOS using Homebrew, so revert back to the use of nameref.

Add support for snapshot tests: for a test named "X", if there is
also a file "X.snap", then run it as a snapshot test.

Create a new test directory "libs" for library tests, and populate it
with Prelude and StmtFSM, as snapshot tests.